+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Login using a Google Account.
 - Add default trigrams, quadgrams, and most common words
 - Highlight errors in the practice session
+- Calculate WPM and average word WPM